Fixed compiling error
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@5850 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ protocolgame.cpp
@@ -538,7 +538,7 @@ void ProtocolGame::parsePacket(NetworkMessage &msg)
 	}
 
 	++m_packetCount;
-	if(m_packetCount > g_config.getNumber(ConfigManager::PACKETS_PER_SECOND))
+	if(m_packetCount > (uint32_t)g_config.getNumber(ConfigManager::PACKETS_PER_SECOND))
 		return;	
 
 	uint8_t recvbyte = msg.get<char>();
